requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
Flask-Migrate
Flask-Mail
mysql-connector-python
requests
matplotlib
numpy
twilio
